feat: setup clerk user management

+++ server/api/songs/[id].delete.ts
@@ -1,9 +1,11 @@
 import { songService } from '../../modules/songs/song.service'
+import { requireUserId } from '../../utils/auth'
 
 export default defineEventHandler((event) => {
+  const userId = requireUserId(event)
   const id = getRouterParam(event, 'id')!
 
-  const song = songService.delete({ id })
+  const song = songService.delete({ id, userId })
   if (!song) {
     throw createError({
       statusCode: 404,
